+++ projects/net.splitcells.os.state.interface/src/main/bash/echo/echo.error.sh
@@ -17,3 +17,14 @@
 
 echo.line.current.clear
 >&2 echo "error" $@
+
+if [ "$echo_error_notifies_user" == 'true' ]
+then
+  # Send a notification to the current user, if the environment is set accordingly.
+  # Thereby, it is more easily to know, if some optional feature was not executed during a long command.
+  # For instance, when a command synchronizes a local repo with all other personal computers,
+  # a notification for each not available computer is helpful.
+  # In this case it can be normal, that some computers were not available during each synchronization,
+  # and therefore the execution as a whole does not have to be aborted or marked as a failure.
+  notify.error $@
+fi